screenlink/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-socketio>=5.9.0
jinja2>=3.1.0
python-multipart>=0.0.6
